requirements.txt
beautifulsoup4
discord
pandas
pyyaml
